# docs/swagger.yaml
basePath: /api
definitions:
  request.ConfirmOtpDto:
    properties:
      code:
        type: integer
      phone:
        type: string
    type: object
  request.RegisterOrLoginDto:
    properties:
      phone:
        type: string
    type: object
  response.UserDto:
    properties:
      createdAt:
        type: string
      id:
        type: string
      phone:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: mehranhhm@gmail.com
    name: API Support
  description: This is a sample User API
  title: User API
  version: "1.0"
paths:
  /auth/confirm-otp:
    post:
      consumes:
      - application/json
      description: Verify OTP code sent to user's phone and complete authentication
      parameters:
      - description: OTP confirmation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ConfirmOtpDto'
      produces:
      - application/json
      responses:
        "200":
          description: Authentication successful
        "400":
          description: Invalid request data
        "401":
          description: Invalid OTP code
        "422":
          description: Validation failed
        "500":
          description: Internal server error
      summary: Confirm OTP code
      tags:
      - auth
  /auth/register-or-login:
    post:
      consumes:
      - application/json
      description: Register a new user or login existing user by phone number. Sends
        OTP for verification.
      parameters:
      - description: Registration/Login data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.RegisterOrLoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent successfully and user already existed
        "201":
          description: OTP sent successfully and reated new user
        "400":
          description: Invalid request data
        "422":
          description: Validation failed
        "500":
          description: Internal server error
      summary: Register or login user
      tags:
      - auth
  /user/:
    get:
      consumes:
      - application/json
      description: Get paginated list of users with optional filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Filter by phone number
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/response.UserDto'
            type: array
        "400":
          description: Invalid query parameters
        "500":
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Get users list
      tags:
      - users
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Get user details by their UUID
      parameters:
      - description: User ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/response.UserDto'
        "400":
          description: Invalid UUID format
        "404":
          description: User not found
        "500":
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"